pyproject.toml
[tool.poetry]
name = "fdadata"
version = "0.1.0"
description = "Project to work on dbt test"
authors = ["Piotr Mierzejewski"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
jupyter = "^1.1.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
